fixed bug when creating packed arrays using more than (2^31)-7 bits (~256MB) of memory

+++ src/main/java/com/dynatrace/hash4j/util/PackedArray.java
@@ -222,7 +222,7 @@ public final class PackedArray {
 
     @Override
     public int numBytes(int length) {
-      return (bitSize * length + 7) >>> 3;
+      return Math.toIntExact(((long) bitSize * length + 7) >>> 3);
     }
 
     @Override
